Bug #30567060 INNOBASE/SRV: REFINE SRV0CONC WITH C++11 ATOMICS
- This commit is based on contribution made by Yibo Cai from ARM.
- The n_waiting and n_active variables have been made std::atomic and the
built_in atomic operations have been removed from srv0conc.

RB#24805 Approved by Jakub Lopuszanski <jakub.lopuszanski@oracle.com>

+++ storage/innobase/handler/ha_innodb.cc
@@ -1601,8 +1601,9 @@ static inline dberr_t innobase_srv_conc_enter_innodb(row_prebuilt_t *prebuilt) {
 
     } else if (trx->mysql_thd != nullptr &&
                thd_is_replication_slave_thread(trx->mysql_thd)) {
-      UT_WAIT_FOR(srv_conc_get_active_threads() < srv_thread_concurrency,
-                  srv_replication_delay * 1000);
+      UT_WAIT_FOR(
+          srv_conc_get_active_threads() < (int32_t)srv_thread_concurrency,
+          srv_replication_delay * 1000);
 
     } else {
       err = srv_conc_enter_innodb(prebuilt);
